requirements.txt
asgiref==3.11.0
astroid==3.3.10
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.2.1
codecov-cli==11.2.6
coverage==7.13.4
crispy-bootstrap5==2023.10
dill==0.4.0
Django==6.0.2
django-appconf==1.2.0
django-bootstrap-modal-forms==3.0.5
django-chartjs==2.3.0
django-crispy-forms==2.1
django-crum==0.7.9
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-select2==8.4.8
django-tables2==2.7.5
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
ijson==3.4.0.post0
iniconfig==2.3.0
isort==6.0.1
lxml==6.0.2
mccabe==0.7.0
mysqlclient==2.2.7
ofxparse==0.21
packaging==25.0
pip-review==1.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
Pygments==2.19.2
pylint==3.3.7
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
responses==0.21.0
sentry-sdk==2.52.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
test_results_parser==0.6.1
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
urllib3==2.6.3
whitenoise==6.9.0
wrapt==1.17.2
